fix: screen recording crash
Fix screen recording crash by disabling faulty encoder. The screen recorder was crashing on Android 16 due to a Failed to create video encoder error. This was caused by the system trying to use the OMX.qcom.video.encoder.hevc.cq encoder, which was not properly configured in the media profiles. This change disables the OMX.qcom.video.encoder.hevc.cq by commenting it out in media_codecs.xml. This forces the screen recorder to fall back to a different, functional encoder, which resolves the crash.

Signed-off-by: Kshitij Bhale <kbhale01@outlook.com>